+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
 - _Scheme less_ urls now keep their leading `//` when converted to strings
 
+### Fixed
+
+- Url parsing idempotency with partially encoded characters
+
 ## 5.3.0 - 2026-05-14
 
 ### Changed