+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - Url parsing idempotency with partially encoded characters
 - `Query::of()` silently lost characters
+- `Path::of()` silently lost characters
 
 ## 5.3.0 - 2026-05-14