+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Url parsing idempotency with partially encoded characters
 - `Query::of()` silently lost characters
 - `Path::of()` silently lost characters
+- `Host::of()` silently lost characters
 
 ## 5.3.0 - 2026-05-14