+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - `Path::of()` silently lost characters
 - `Host::of()` silently lost characters
 - `Password::of()` silently lost characters
+- `User::of()` silently lost characters
 
 ## 5.3.0 - 2026-05-14